+++ src/edu/sjsu/cmpe297/db/object/Company.java
@@ -3,16 +3,16 @@ package edu.sjsu.cmpe297.db.object;
 public class Company implements DatabaseObject {
 
 	// member variables
-	private Integer facebookId; //TODO: choose Long because there are a lot of facebook ids out there
+	private Long facebookId; //TODO: choose Long because there are a lot of facebook ids out there
 	private String name;
 	
-	public Company(Integer facebookId, String name) {
+	public Company(Long facebookId, String name) {
 		this.facebookId = facebookId;
 		this.name = name;
 	}
 	
 	// accessor methods
-	public Integer getFacebookId() {
+	public Long getFacebookId() {
 		return facebookId;
 	}
 	
@@ -20,7 +20,7 @@ public class Company implements DatabaseObject {
 		return name;
 	}
 	
-	public void setFacebookId(Integer facebookId) {
+	public void setFacebookId(Long facebookId) {
 		this.facebookId = facebookId;
 	}
