+++ src/edu/sjsu/cmpe297/db/object/Users.java
@@ -3,16 +3,16 @@ package edu.sjsu.cmpe297.db.object;
 public class Users implements DatabaseObject {
 
 	// member variables
-	private Integer facebookId; //TODO: choose Long because there are a lot of facebook ids out there
+	private Long facebookId; //TODO: choose Long because there are a lot of facebook ids out there
 	private String name;
 	
-	public Users(Integer facebookId, String name) {
+	public Users(Long facebookId, String name) {
 		this.facebookId = facebookId;
 		this.name = name;
 	}
 	
 	// accessor methods
-	public Integer getFacebookId() {
+	public Long getFacebookId() {
 		return facebookId;
 	}
 	
@@ -20,7 +20,7 @@ public class Users implements DatabaseObject {
 		return name;
 	}
 	
-	public void setFacebookId(Integer facebookId) {
+	public void setFacebookId(Long facebookId) {
 		this.facebookId = facebookId;
 	}
